+++ api/topology/src/main/java/quarks/topology/plumbing/Valve.java
@@ -1,4 +1,4 @@
-package quarks.analytics.sensors;
+package quarks.topology.plumbing;
 
 import java.util.Objects;
